front_back_separation/25-11-2022/LesGaragistesDuSud/requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.2
django-cors-headers==3.13.0
django-extensions==3.2.1
django-stubs==1.12.0
django-stubs-ext==0.5.0
djangorestframework==3.14.0
djangorestframework-stubs==1.7.0
Faker==15.3.3
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy==0.982
mypy-extensions==0.4.3
pycodestyle==2.9.1
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.5
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
toposort==1.7
types-Markdown==3.4.2.1
types-pytz==2022.5.0.0
types-PyYAML==6.0.12
types-requests==2.28.11.2
types-urllib3==1.26.25.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
